// gen/esi.ts
export namespace esi {
    export namespace alliance {
        /**
         * This is the response type for the route, [`GET /v2/alliances/{alliance_id}/`](https://esi.evetech.net//#!/Alliance/get_alliances_alliance_id).
         */
        export interface Alliance {
            /**
             * The full name of the alliance.
             */
            alliance_name: string;
            date_founded: string;
            /**
             * The executor corporation ID, if this alliance is not closed.
             */
            executor_corp?: number;
            /**
             * The short name of the alliance.
             */
            ticker: string;
        }
        /**
         * This is the response type for the route, [`GET /v1/alliances/{alliance_id}/icons/`](https://esi.evetech.net//#!/Alliance/get_alliances_alliance_id_icons).
         */
        export interface Icons {
            px128x128?: string;
            px64x64?: string;
        }
        export interface Name {
            alliance_id: number;
            alliance_name: string;
        }
    }
    export namespace character {
// ...
        export interface Search {
            agent?: number[];
            alliance?: number[];
            character?: number[];
            constellation?: number[];
            corporation?: number[];
            faction?: number[];
            inventorytype?: number[];
            region?: number[];
            solarsystem?: number[];
            station?: number[];
            structure?: number[];
            wormhole?: number[];
        }
// ...
    }
// ...
    export interface Status {
        /**
         * Current online player count.
         */
        players: number;
        /**
         * Running version as string.
         */
        server_version: string;
        /**
         * Server start timestamp.
         */
        start_time: string;
        /**
         * If the server is in VIP mode.
         */
        vip?: boolean;
    }
// ...
}
// ...
export interface Responses {
// ...
    get_search: esi.Search;
// ...
    get_status: esi.Status;
// ...
    post_universe_names: esi.universe.Name[];
// ...
}